+++ AppSolidarIT_Angular.Server/Controllers/ThemeController.cs
@@ -1,6 +1,7 @@
 ﻿using AppSolidarIT_Angular.Server.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Data;
 
 namespace AppSolidarIT_Angular.Server.Controllers
 {
@@ -17,6 +18,8 @@ namespace AppSolidarIT_Angular.Server.Controllers
         }
 
 
+
+
         [HttpGet()]
         public IEnumerable<Theme> GetAllThemes()
         {
@@ -24,36 +27,86 @@ namespace AppSolidarIT_Angular.Server.Controllers
 
         }
 
-        [HttpPost]
-        public void CreateTheme(Theme theme)
-        {
+        [HttpGet("{Id}")]
 
-            _context.Themes.Add(theme);
-            _context.SaveChanges();
+        public ActionResult<Theme> GetTheme(int Id)
+        {
+            var theme = _context.Themes?.Find(Id);
+            if (theme == null)
+            {
+                return NotFound();
+            }
+            return Ok(theme);
+
+        }
+
+        [HttpPost]
+        public ActionResult CreateTheme([Bind(include: "Name, DesccriptionShort")] Theme theme)
+        {
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    _context.Themes.Add(theme);
+                    _context.SaveChanges();
+                }
+
+
+            }
+            catch (DataException)
+            {
+                ModelState.AddModelError("", "Impossible de créer le produit, verifier les conditions et essayez encore");
+            }
+            return Ok(theme);
         }
 
         [HttpPut]
-        public void UpdateProduct(Theme theme)
+        public ActionResult UpdateProduct([Bind(include: "Name, DesccriptionShort")] Theme theme)
         {
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    _context.Themes.Remove(theme);
+                    _context.SaveChanges();
+                }
+            }
+            catch (DataException)
+            {
+                ModelState.AddModelError("", "Impossible de mettre à jour le theme, verifier les conditions et essayez encore");
+            }
+            return Ok(theme);
 
-            _context.Themes.Remove(theme);
-            _context.SaveChanges();
+
         }
 
 
         [HttpDelete]
 
-        public void DeleteProduct(Theme theme)
+        public ActionResult DeleteProduct(Theme theme)
         {
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    _context.Themes.Remove(theme);
+                    _context.SaveChanges();
+                }
+            }
+            catch (DataException)
+            {
 
-           _context.Themes.Remove(theme);
-            _context.SaveChanges();
+                ModelState.AddModelError("", "Impossible de supprimer le theme");
+            }
+            return Ok(theme);
+
+
         }
     }
 
-  
 
 
-        
-        
+
+
+
 }
